Introduced routines for MPI shared memory
MPI shared memory is a concept for allocating memory accessibly by ranks
that are connected to the same physical memory, i.e. on the same node in
a cluster. In this basic implementation here, we can use this concept to
reduce the memory usage of an application if large arrays are stored
only once per node and not once per MPI task.

See test/test_shared_memory.f90 for a short example.